Do not use carwriter if using separate datastore for ad data

+++ internal/ingest/linksystem.go
@@ -363,7 +363,7 @@ func (ing *Ingester) ingestAd(publisherID peer.ID, adCid cid.Cid, ad schema.Adve
 		gatherCids := func(_ peer.ID, c cid.Cid, _ dagsync.SegmentSyncActions) {
 			hamtCids = append(hamtCids, c)
 		}
-		if ing.carWriter == nil {
+		if ing.carWriter == nil && ing.dsAds == ing.ds {
 			defer func() {
 				for _, c := range hamtCids {
 					err := ing.dsAds.Delete(ctx, datastore.NewKey(c.String()))
@@ -530,7 +530,7 @@ func (ing *Ingester) ingestAd(publisherID peer.ID, adCid cid.Cid, ad schema.Adve
 // operation. This function is used as a scoped block hook, and is called for
 // each block that is received.
 func (ing *Ingester) ingestEntryChunk(ctx context.Context, ad schema.Advertisement, entryChunkCid cid.Cid, chunk schema.EntryChunk, log *zap.SugaredLogger) error {
-	if ing.carWriter == nil {
+	if ing.carWriter == nil && ing.dsAds == ing.ds {
 		defer func() {
 			// Remove the content block from the data store now that processing it
 			// has finished. This prevents storing redundant information in several
